Breaking: Remove LOG_CACHE_SKIP_AUTH env var option
Removes the LOG_CACHE_SKIP_AUTH env var as an option for telling the
plugin not to use authentication headers in requests. This is a relic of
when the plugin had a dual function as a standalone CLI, and is no
longer necessary since Log Cache should always require authentication
headers, which can be easily fetched with cf CLI connection methods.